Move RebindSafeMBeanServer to presto-plugin-toolkit
