+++ presto-cassandra/src/main/java/io/prestosql/plugin/cassandra/CassandraConnectorFactory.java
@@ -13,11 +13,10 @@
  */
 package io.prestosql.plugin.cassandra;
 
-import com.google.inject.Binder;
 import com.google.inject.Injector;
-import com.google.inject.Module;
 import io.airlift.bootstrap.Bootstrap;
 import io.airlift.json.JsonModule;
+import io.prestosql.plugin.base.jmx.RebindSafeMBeanServer;
 import io.prestosql.spi.connector.Connector;
 import io.prestosql.spi.connector.ConnectorContext;
 import io.prestosql.spi.connector.ConnectorFactory;
