+++ presto-plugin-toolkit/src/main/java/io/prestosql/plugin/base/jmx/RebindSafeMBeanServer.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.plugin.hive;
+package io.prestosql.plugin.base.jmx;
 
 import io.airlift.log.Logger;
