+++ presto-raptor-legacy/src/main/java/io/prestosql/plugin/raptor/legacy/RaptorConnectorFactory.java
@@ -18,10 +18,10 @@ import com.google.inject.Injector;
 import com.google.inject.Module;
 import io.airlift.bootstrap.Bootstrap;
 import io.airlift.json.JsonModule;
+import io.prestosql.plugin.base.jmx.RebindSafeMBeanServer;
 import io.prestosql.plugin.raptor.legacy.backup.BackupModule;
 import io.prestosql.plugin.raptor.legacy.security.RaptorSecurityModule;
 import io.prestosql.plugin.raptor.legacy.storage.StorageModule;
-import io.prestosql.plugin.raptor.legacy.util.RebindSafeMBeanServer;
 import io.prestosql.spi.NodeManager;
 import io.prestosql.spi.PageSorter;
 import io.prestosql.spi.connector.Connector;
