+++ presto-thrift/pom.xml
@@ -19,6 +19,11 @@
     <dependencies>
         <dependency>
             <groupId>io.prestosql</groupId>
+            <artifactId>presto-plugin-toolkit</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>io.prestosql</groupId>
             <artifactId>presto-thrift-api</artifactId>
         </dependency>
